Was able to deal with the share function from user library to collective library
