+++ src/App.jsx
@@ -1,18 +1,20 @@
 import { Route, Routes } from "react-router-dom";
 import HomePage from "./components/Homepage";
-import Signup from "./components/Signup";
-import Login from "./components/Login";
-import Profile from "./components/Profile";
+import Signup from "./components/User/Signup";
+import Login from "./components/User/Login";
+import Profile from "./components/User/Profile";
 import Outlet from "./components/Outlet";
-import IsPrivate from "./components/IsPrivate";
-import CreateWebDesigns from "./components/CreateWebDesigns";
-import CreateFonts from "./components/CreateFonts";
-import CreateImages from "./components/CreateImages";
-import CollectiveImageLibrary from "./components/CollectiveImageLibrary";
-import CollectiveFontLibrary from "./components/CollectiveFontLibrary";
-import CollectiveWebDesignLibrary from "./components/CollectiveWebDesignLibrary";
+import IsPrivate from "./components/User/IsPrivate";
+import CreateWebDesigns from "./components/Create/CreateWebDesigns";
+import CreateFonts from "./components/Create/CreateFonts";
+import CreateImages from "./components/Create/CreateImages";
+import CollectiveImageLibrary from "./components/CollectiveLibrary/CollectiveImageLibrary";
+import CollectiveFontLibrary from "./components/CollectiveLibrary/CollectiveFontLibrary";
+import CollectiveWebDesignLibrary from "./components/CollectiveLibrary/CollectiveWebDesignLibrary";
 import NotFound from "./components/NotFound";
-// import WebDesignCardDetails from "./components/WebDesignCardDetails";
+import WebDesignCardDetails from "./components/ReadForUser/WebDesignCardDetails";
+import ImageCardDetails from "./components/ReadForUser/ImageCardDetails";
+import FontCardDetails from "./components/ReadForUser/FontCardDetails";
 
 function App() {
   return (
@@ -24,10 +26,25 @@ function App() {
         <Route path="/createWebdesigns" element={<CreateWebDesigns />} />
         <Route path="/createFonts" element={<CreateFonts />} />
         <Route path="/createImages" element={<CreateImages />} />
-        <Route path="/collectiveImageLibrary" element={<CollectiveImageLibrary />} />
-        <Route path="/collectiveFontLibrary" element={<CollectiveFontLibrary />} />
-        <Route path="/collectiveWebDesignLibrary" element={<CollectiveWebDesignLibrary />} />
-        {/* <Route path="/web-design/:id" element={<WebDesignCardDetails webD />} /> */}
+        <Route
+          path="/collectiveImageLibrary"
+          element={<CollectiveImageLibrary />}
+        />
+        <Route
+          path="/collectiveFontLibrary"
+          element={<CollectiveFontLibrary />}
+        />
+        <Route
+          path="/collectiveWebDesignLibrary"
+          element={<CollectiveWebDesignLibrary />}
+        />
+
+        <Route
+          path="/web-design/:webDesignId"
+          element={<WebDesignCardDetails />}
+        />
+        <Route path="/web-design/:imageId" element={<ImageCardDetails />} />
+        <Route path="/web-design/:fontId" element={<FontCardDetails />} />
 
         <Route
           path="/profile"
