+++ src/components/Create/CreateWebDesigns.jsx
@@ -1,17 +1,18 @@
+import { useContext } from "react";
 import { useState } from "react";
 import { useNavigate } from "react-router-dom";
 // import { UserContext } from "../contexts/UserContext";
 // import axios from "axios";
-import service from "../services/file-image-upload.service";
+import service from "../../services/file-upload.service";
 import { hourglass } from "ldrs";
 
 hourglass.register();
 
 // Default values shown
 
-export default function CreateImages() {
+export default function CreateWebDesigns() {
   const [name, setName] = useState("");
-  const [contributor, setContributor] = useState("");
+  const [description, setDescription] = useState("");
   const [imageUrl, setImageUrl] = useState("");
   const [loading, setLoading] = useState(false);
   const navigate = useNavigate();
@@ -26,7 +27,7 @@ export default function CreateImages() {
       // req.body to .create() method when creating a new movie in '/api/movies' POST route
       uploadData.append("imageUrl", e.target.files[0]);
 
-      const response = await service.uploadImage(uploadData);
+      const response = await service.uploadWebDesign(uploadData);
 
       // response carries "fileUrl" which we can use to update the state
       setImageUrl(response.fileUrl);
@@ -40,21 +41,21 @@ export default function CreateImages() {
     e.preventDefault();
 
     try {
-      const res = await service.createImage({
+      const res = await service.createWebDesign({
         name,
-        contributor,
+        description,
         imageUrl,
       });
 
       // Reset the form
       setName("");
-      setContributor("");
+      setDescription("");
       setImageUrl("");
 
       // navigate to another page
       navigate("/profile");
     } catch (error) {
-      console.error("Error while adding the new image: ", error);
+      console.error("Error while adding the new web design: ", error);
     }
   };
 
@@ -102,9 +103,9 @@ export default function CreateImages() {
     );
   } else {
     return (
-      <form onSubmit={handleSubmit}>
-        <label>
-          Image Name:
+      <form onSubmit={handleSubmit} className="max-w-lg mx-auto">
+        <label className="block mb-2 text-sm font-medium text-gray-900 ">
+          Web Design Name:
           <input
             name="name"
             value={name}
@@ -112,26 +113,34 @@ export default function CreateImages() {
             onChange={(e) => {
               setName(e.target.value);
             }}
+            className="block w-full p-2 text-gray-900 border border-gray-300 rounded-lg bg-gray-50 sm:text-xs focus:ring-blue-500 focus:border-blue-500"
           />
         </label>
 
-        <label>
-          Image contributor:
+        <label className="block mb-2 text-sm font-medium text-gray-900">
+          Web Design Description:
           <input
-            name="contributor"
-            value={contributor}
+            name="description"
+            value={description}
             type="text"
             onChange={(e) => {
-              setContributor(e.target.value);
+              setDescription(e.target.value);
             }}
+            className="block w-full p-4 text-gray-900 border border-gray-300 rounded-lg bg-gray-50 sm:text-md focus:ring-blue-500 focus:border-blue-500"
           />
         </label>
 
-        <label>
-          Image:
-          <input type="file" onChange={(e) => handleFileUpload(e)} />
+        <label className="block mb-2 text-sm font-medium text-gray-900">
+          Web Design Image:
+          <input
+            type="file"
+            onChange={(e) => handleFileUpload(e)}
+            className="block focus:outline-none w-full text-sm text-gray-900 border border-gray-300 rounded-lg cursor-pointer bg-gray-50 "
+          />
         </label>
-        <button className="">Create Image</button>
+        <button className="relative my-4 flex justify-center mx-auto scale-100 text-xl overflow-hidden rounded-lg bg-gradient-to-br from-indigo-600 from-40% to-indigo-400 px-4 py-2 font-semibold text-gray-100 transition-transform hover:scale-105 active:scale-95">
+          Create Web Design
+        </button>
       </form>
     );
   }
