+++ src/components/Layout/Header.jsx
@@ -1,8 +1,7 @@
 import { FiMenu, FiArrowUpRight } from "react-icons/fi";
-import { useContext, useEffect, useRef, useState } from "react";
+import { useEffect, useRef, useState } from "react";
 import { useAnimate, motion } from "framer-motion";
-import evoke from "../../public//Evoke.png";
-import { UserContext } from "../contexts/UserContext";
+import evoke from "../../../public//Evoke.png";
 
 const Header = () => {
   return (
@@ -150,16 +149,40 @@ const SignUpButton = () => {
 };
 const DashboardButton = () => {
   return (
-    <button className="group relative font-semibold text-xl scale-100 overflow-hidden rounded-lg px-4 py-2 transition-transform hover:scale-105 active:scale-95">
-      <a
-        href="/collectiveLibrary"
-        className="relative z-10 text-gray/90 transition-colors group-hover:text-gray"
-      >
-        {/* change a ref to dashboard */}
-        Our Library
-      </a>
-      <span className="absolute inset-0 z-0 bg-gradient-to-br from-gray/20 to-gray/5 opacity-0 transition-opacity group-hover:opacity-100" />
-    </button>
+    <>
+      <button className="group relative font-semibold text-xl scale-100 overflow-hidden rounded-lg px-4 py-2 transition-transform hover:scale-105 active:scale-95">
+        <a
+          href="/collectiveWebDesignLibrary"
+          className="relative z-10 text-gray/90 transition-colors group-hover:text-gray"
+        >
+          {/* change a ref to dashboard */}
+          Web-Design Library
+        </a>
+        <span className="absolute inset-0 z-0 bg-gradient-to-br from-gray/20 to-gray/5 opacity-0 transition-opacity group-hover:opacity-100" />
+      </button>
+
+      <button className="group relative font-semibold text-xl scale-100 overflow-hidden rounded-lg px-4 py-2 transition-transform hover:scale-105 active:scale-95">
+        <a
+          href="/collectiveImageLibrary"
+          className="relative z-10 text-gray/90 transition-colors group-hover:text-gray"
+        >
+          {/* change a ref to dashboard */}
+          Image Library
+        </a>
+        <span className="absolute inset-0 z-0 bg-gradient-to-br from-gray/20 to-gray/5 opacity-0 transition-opacity group-hover:opacity-100" />
+      </button>
+
+      <button className="group relative font-semibold text-xl scale-100 overflow-hidden rounded-lg px-4 py-2 transition-transform hover:scale-105 active:scale-95">
+        <a
+          href="/collectiveFontLibrary"
+          className="relative z-10 text-gray/90 transition-colors group-hover:text-gray"
+        >
+          {/* change a ref to dashboard */}
+          Font Library
+        </a>
+        <span className="absolute inset-0 z-0 bg-gradient-to-br from-gray/20 to-gray/5 opacity-0 transition-opacity group-hover:opacity-100" />
+      </button>
+    </>
   );
 };
