+++ src/components/User/IsPrivate.jsx
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { UserContext } from "../contexts/UserContext";
+import { UserContext } from "../../contexts/UserContext";
 import { Navigate } from "react-router-dom";
 
 function IsPrivate({ children }) {
