+++ src/components/User/Profile.jsx
@@ -1,14 +1,17 @@
 import { useContext } from "react";
-import { UserContext } from "../contexts/UserContext";
+import { UserContext } from "../../contexts/UserContext";
 import { Link } from "react-router-dom";
-import ProfileHeader from "./ProfileComponents/ProfileHeader";
-import WebDesignList from "./WebDesignList";
+import ProfileHeader from "../ProfileComponents/ProfileHeader";
+import WebDesignList from "../List/WebDesignList";
 
 export default function Profile() {
   const { user, handleLogout } = useContext(UserContext);
   return (
     <div>
       <ProfileHeader />
+      <button onClick={handleLogout} className="m-4">
+        Logout
+      </button>
       <div className="p-20">
         <Link className="px-4" to="/createWebdesigns">
           Create Web Design
@@ -21,9 +24,6 @@ export default function Profile() {
         </Link>
       </div>
       <WebDesignList />
-      <button onClick={handleLogout} className="p-20">
-        Logout
-      </button>
     </div>
   );
 }
