+++ src/components/User/Signup.jsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
 import axios from "axios";
-import evoke from "../../public/Evoke.png";
+import evoke from "../../../public/Evoke.png";
 import { motion } from "framer-motion";
 
 const API_URL = "http://localhost:5005";
