--- src/components/WebDesignList.jsx
@@ -1,92 +1,0 @@
-import { useEffect, useState } from "react";
-import service from "../services/file-upload.service";
-import { Link } from "react-router-dom";
-import axios from "axios";
-// import WebDesignCardDetails from "./WebDesignCardDetails";
-
-export default function WebDesignList() {
-  // const { webDesigns, setWebDesigns, getWebDesigns } =
-  //   useContext(WebDesignContext);
-
-  const [webDesigns, setWebDesigns] = useState([]);
-
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const data = await service.getWebDesigns();
-        setWebDesigns(data);
-      } catch (error) {
-        console.error(error);
-      }
-    };
-
-    fetchData();
-  }, []);
-
-  // const handleDone = (webDesignId) => {
-  //   const mappedWebDesigns = webDesigns.map((elem) => {
-  //     if (elem.id === webDesignId) {
-  //       elem.done = !elem.done;
-  //     }
-  //     return elem;
-  //   });
-  //   setWebDesigns(mappedWebDesigns);
-  // };
-
-  const handleDelete = async (id) => {
-    try {
-      //make an axios call to the back to delete the todo as well
-      const { data } = await axios.delete(
-        `http://localhost:5005/web-design/${id}`
-      );
-      console.log("design was deleted ", data);
-    } catch (err) {
-      console.log("there was an error deleting", err);
-    }
-
-    // perfect for the DOM
-    const filteredWebDesigns = webDesigns.filter((curr) => {
-      if (curr._id !== id) {
-        return true;
-      }
-    });
-    setWebDesigns(filteredWebDesigns);
-  };
-
-  return (
-    <div className="flex justify-between">
-      <div className="grid grid-cols-4 gap-5">
-        {webDesigns &&
-          webDesigns.map((webDesign) => {
-            return (
-              <div key={webDesign._id}>
-                <Link to={`/web-design/${webDesign._id}`}>
-                  <img
-                    src={webDesign.imageUrl}
-                    alt={webDesign.name}
-                    style={{
-                      objectFit: "cover",
-                      maxWidth: "100%",
-                      height: "100%",
-                    }}
-                  />
-                </Link>
-                {/* <WebDesignCardDetails
-                  key={webDesign.id}
-                  webDesigns={webDesigns}
-                /> */}
-                {/* <button
-                  onClick={() => {
-                    handleDelete(webDesign._id);
-                  }}
-                  className="py-4"
-                >
-                  Delete
-                </button> */}
-              </div>
-            );
-          })}
-      </div>
-    </div>
-  );
-}
